Add new model. See #26.
