+++ files/rsum/rsum/home/models.py
@@ -7,6 +7,7 @@ import json
 import yaml
 
 from django.db import models
+from home.schema.entryitem import EntryItem
 
 sections_list = [
     'intro',
@@ -19,244 +20,3 @@ sections_list = [
 ]
 
 
-
-
-    
-class SubSection(models.Model):
-    section = models.ForeignKey(Section, on_delete=models.CASCADE)
-    name = models.CharField(max_length=200, null=True)
-    value = models.CharField(max_length=200, null=True)
-
-    def get_sub_section(self, section):
-        subsections = []
-        for subsection in list(
-            SubSection.objects.filter(
-                section = section
-            ).values() 
-        ):
-            p = Project() 
-            if (
-                subsection.get('value') == u"<type 'list'>" or
-                subsection.get('value') == u"<type 'dict'>"
-            ):
-                subsection.update({
-                    'value': p.get_projects(
-                        SubSection.objects.filter(
-                            id = subsection.get('id')
-                        )
-                    )
-                })
-            else:
-                subsection.update({
-                    'value': list(
-                        Project.objects.filter(
-                            sub_section = subsection.get('id')  
-                        ).values()
-                    )
-                })
-            subsections.append(subsection)
-        return subsections
-                    
-    
-    def save_sub_sections(self, sub_section, section):
-        projects = []
-        if type(sub_section) == type(dict()):
-            for k, v in sub_section.iteritems():
-                ss_i = SubSection()
-                ss_i.section = section
-                ss_i.value = type(sub_section)
-                ss_i.name = k
-                ss_i.save()
-                p = Project() 
-                projects.append(p.save_projects(v, ss_i))
-            return projects
-
-        if type(sub_section) == type(list()):
-            for c,i in enumerate(sub_section):
-                ss_i = SubSection()
-                ss_i.section = section
-                ss_i.value = type(sub_section)
-                ss_i.name = str(c) 
-                ss_i.save()
-                p = Project()
-                projects.append(p.save_projects(i, ss_i))
-            return projects 
-            
-
-class Project(models.Model):
-    sub_section = models.ForeignKey(SubSection, on_delete=models.CASCADE)
-    name = models.CharField(max_length=200, null=True)
-    value = models.CharField(max_length=200, null=True)
-
-    def get_projects(self, subsection):
-        projects = []
-        for project in list(
-            Project.objects.filter(
-                sub_section = subsection
-            ).values()
-        ):
-            pli = ProjectItems()
-            if project.get('value') == u"<type 'dict'>":
-                project.update({
-                    'value': pli.get_project_items(
-                        Project.objects.filter(
-                            id = project.get('id')
-                        )
-                    ) 
-                })
-                
-            projects.append(project) 
-        return projects
-
-    def save_projects(self, projects, sub_section):
-        if type(projects) == type(dict()):
-            for k, v in projects.iteritems():
-                p_i = Project()
-                p_i.sub_section = sub_section
-                p_i.name = k 
-                if type(v) == type(dict()):
-                    p_i.value = type(v)
-                    p_i.save()
-                    pi = ProjectItems()
-                    pi.save_project_items(v, p_i)
-                else:
-                    p_i.value = v
-                    p_i.save()
-            return Project.objects.values_list() 
-
-        if type(projects) == type(list()):
-            for k,v in enumerate(projects):
-                p_i = Project()
-                p_i.sub_section = sub_section
-                p_i.name = k
-                p_i.value = type(v)
-                p_i.save()
-                pi = ProjectItems()
-                pi.save_project_items(v, p_i)
-            return Project.objects.values_list()
-
-        if type(projects) == type(str()):
-            p_i = Project()
-            p_i.sub_section = sub_section
-            p_i.value = projects
-            p_i.save()
-
-
-class ProjectItems(models.Model):
-    project = models.ForeignKey(Project, on_delete=models.CASCADE)
-    name = models.CharField(max_length=200)
-    value = models.CharField(max_length=200, null = True) 
-    iterable = models.BooleanField(default=False)
-
-    def get_project_items(self, project):
-        project_items = []
-        for project_item in list(
-            ProjectItems.objects.filter(
-                project = project
-            ).values()
-        ):
-            if project_item.get('value') == u"<type 'dict'>":
-                pe = ProjectEntry()
-                project_item.update({
-                    'value': pe.get_entry(
-                        ProjectItems.objects.filter(
-                            id = project_item.get('id')
-                        )
-                    )
-                })        
-                project_items.append(project_item)
-            else:
-                project_items.append(project_item)
-        return project_items
-
-    def save_project_items(self, project_item, project):
-        if type(project_item) == type(dict()):
-            for key, p_entry in project_item.iteritems():
-                pi_i = ProjectItems()
-                pi_i.project = project
-                pi_i.name = key
-                if type(p_entry) == type(dict()):
-                    pi_i.value = type(p_entry)
-                    pi_i.iterable = True
-                    pi_i.save()
-                    pe = ProjectEntry()
-                    pe.save_entry(p_entry, pi_i)
-                else:
-                    pi_i.value = p_entry
-                    pi_i.save()
-            return ProjectItems.objects.values_list() 
-
-        if type(project_item) == type(str()):
-            pi_i = ProjectItems()
-            pi_i.project = project
-            pi_i.name = "<type 'str'>"
-            pi_i.save()
-        return pi_i
-
-
-class ProjectEntry(models.Model):
-    project_item_list = models.ForeignKey(ProjectItems, on_delete=models.CASCADE)
-    name = models.CharField(max_length=200, null = True)
-    value = models.CharField(max_length=200, null = True) 
-
-    def get_entry(self, project_item):
-        entries = []
-        for entry in list(
-            ProjectEntry.objects.filter(
-                project_item_list = project_item
-            ).values()
-        ):
-            eli = EntryListItem()
-            entry.update({
-                'value': eli.get_list_item(
-                    ProjectEntry.objects.filter(
-                        id = entry.get('id') 
-                    )
-                ) 
-            })
-            entries.append(entry)
-        return entries
-
-    def save_entry(self, entry, pi_l):
-        if type(entry) == type(dict()):
-            for k, v in entry.iteritems():
-                pe_i = ProjectEntry()
-                pe_i.project_item_list = pi_l
-                pe_i.name = k
-                pe_i.value = type(v)
-                pe_i.save()
-                eli = EntryListItem()
-                eli.save_list_item(v, pe_i)
-            return pe_i 
-
-
-class EntryListItem(models.Model):
-    project_entry = models.ForeignKey(ProjectEntry, on_delete=models.CASCADE)
-    value = models.CharField(max_length=200, null=True)
-    
-    def get_list_item(self, entry):
-        items = []
-        for item in list(
-            EntryListItem.objects.filter(
-                project_entry = entry
-            ).values()
-        ):
-            items.append(item)
-        return items
-            
-
-    def save_list_item(self, list_item, pe):
-        if type(list_item) == type(list()):
-            for i in list_item:
-                eli = EntryListItem()
-                eli.project_entry = pe
-                eli.value = i
-                eli.save()
-            return eli 
-
-        if type(list_item) == type(str()):
-            eli = EntryListItem()
-            eli.project_entry = pe
-            eli.value = list_item
-            eli.save()
-        return eli
